crash fix for I420ToRGB565
To reproduce with the unittest (not included in this CL):

out/Release/libyuv_unittest --gtest_filter=*Convert*I*To* --libyuv_width=1280 --libyuv_height=720

[ RUN      ] LibYUVConvertTest.I420ToRGB24_Opt
[       OK ] LibYUVConvertTest.I420ToRGB24_Opt (585 ms)
[ RUN      ] LibYUVConvertTest.I420ToRGB565_Any
/Users/fbarchard/bin/runyuv10: line 2:  9424 Segmentation fault: 11

Happens when running with Intel with AVX2 on odd width image.
Odd widths are uncommon, and AVX2 is not common on mobile, so doesnt occur much.
This function is used by stagefright.
The change calls a function with a larger alignment value, indicating the AVX2 should do
multiple of 16 pixels with AVX2, and use C code to handle the remainder.  Previously the
function was passed too small of an alignment value, so the AVX2 would overwrite memory.
e.g if it were called with 9 pixels, it would convert 16.

Bug: 37328100
Test: out/Release/libyuv_unittest --gtest_filter=*Convert*I*To* --libyuv_width=1280 --libyuv_height=720
Change-Id: If56113527084256e2c2092938bf573722b16f0df