fix for clang compile error on mips sgtu instruction.
In NDK r14 the assenbler built into llvm fails on sgtu pseudo-op
instruction.  r14 uses clang 3.8.  sgtu is set great than unsigned.

The instruction has been added in 3.9 and 4.0 for future updates
https://reviews.llvm.org/D22752#765a2da5

But the its just a pseudo op for sltu: set less than unsigned.
So this change switches the order of the 2 operands and uses sltu
instead of sgtu.

Bug: 37671160
Test: lunch aosp_mips-eng and build is successful.
Change-Id: I04b5f756787a468d45c85546f2028360130bd6c1